frontend: Release & Acquire buffer instead of preserving it
At the moment frontend.game saves the current *Buffer while the player
is in game and restores it when they leave the game. For every player in
game we remove a *Buffer from the pool forcing more to be created. For
1,000 players we would have 1,000 idle buffers.

This change releases the *Buffer back to the pool when a player enters
the game and acquires a new one when they exit the game.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>