attr: Replace Inventory Add/Remove calls to new Move method
A common pattern in commands is to remove a Thing from one Inventory and
add it to another Inventory. For example moving from one location to
another. However both Add and Remove methods require a number of finders
to be called: FindNarrative, FindPlayer and FindLocate.

This change add a new Inventory.Move method. Move will remove a Thing
from the receiver Inventory and add it to the passed Inventory only
calling each of the finders once. In addition the receiver may be an
*Inventory typed nil and the passed Inventory may be nil. This has the
interesting property of being able to perform the equivelent of just an
Add - receiver is a *Inventory typed nil - or just a Remove - the passed
inventory is nil. This change therefore also amends Inventory.Add and
Inventory.Remove methods to be calls Inventory.Move to avoid code
duplication.

The Add, Remove and Move methods now return true on success and false
otherwise. Commands using Add and Remove have been updated for the new
returned value but have not been changed to use the new Move method yet.

It is interesting to note that the move method uses goto statements
rather than long if statements for clarity.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>