attr,has: Record initial starting inventory in Locate
This change adds an origin field to the Locate struct. Its purpose is to
record the initial starting Inventory of a Thing when it is first placed
into the game world. When a Thing is reset it will be placed at its
origin. The origin can be set via Locate.SetOrigin and retrieved via
Locate.Origin.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>