cmd: Initial JUNK command
The JUNK command is used to dispose of an item a player no longer wants.
After a period of time the item should be reset and placed back into the
world.

This is an initial attempt at a JUNK command so that object resetting
can be tested and worked on. It contains a lot of code that will
eventually be moved into a Reset attribute so that a Thing can be reset
in a number of ways - junking, selling, destroying, clean up when left
lying around etc.

The command checks for vetoes recursively within containers. If an item
in a container cannot be junked the main item cannot be junked. This
behaviour may be modified to just leave the non-junkable items behind.

The command also resets objects recursively within containers. That is,
if we have a ball in a chest and junk the chest the ball and chest will
both be junked and reset.

At this time we cannot handle resetting of pre-populated containers.
That is, if we populate the world with a chest containing a ball the two
will reset independently of each other. What should happen is that the
chest should wait for the ball to reset, and then the chest should
reset. Generally speaking a container should not reset until all of its
content is ready to reset as well.

Currently items are hardcoded to reset after 1 minute with a 30 second
jitter.

A small leather pouch has been added to the Zinara zone for testing
containers within containers. It is a simple container that can have
items placed inside it and it can be placed inside the chest. The
lattice has also been updated with a JUNK veto for testing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>