record: Lowercase data in time.ParseDuration
When calling decoder.Duration the data can be uppercased if being used
on data from a KeyedXxx type method already. However duration specifiers
such h,m and s need to be lowercased for time.ParseDuration. Instead of
sprinkling strings.ToLower when handling this case it is now done in
decoder.Duration itself.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>