attr,has: Inventory Add/Remove/Move return (maybe copy of) has.Thing
The Inventory Add, Remove and Move methods now return the Thing actually
added, removed or moved which may not be the Thing originally passed to
the method. This allows us to surreptitiously replace the Thing being
handled with a copy.

The immediate usefulness of this is in handling things that respawn.
When something is picked up and is set to respawn we can hand out a copy
that does not respawn while removing (hiding) and respawning the
original. By never handling the original item directly it will solve a
lot of issues with containers that may themselves contain respawning or
resetting items.

If a Thing is passed to Add, Remove or Move and is then referenced later
on it is important to use the Thing that was passed back. For example:

  if what = where.Remove(what); what == nil {
    // handle failure
  }

  // what is referenced after Remove

If the Thing is not referenced after it is added, removed or moved it is
sufficient to just check for nil indicating failure:

  if where.Remove(what) == nil {
    // handle failure
  }

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>