attr,has: Add a way of releasing references & resources in attributes
This change adds a Free method to attributes. Attributes can contain a
lot of references to other Thing, Attributes and events. The Free method
can be implemented by any attribute to enable an easy mechanism for
releasing references and resources.

If an attribute does implement its own Free method it should also call
Attribute.Free which currently only sets the parent reference to nil but
may be extended in the future.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>