cmd: DUMP command now accepts a memory address - with BIG caveats!
The dump command can now take an arbitrary memory address in the form
0xc42010f770. The address MUST point to a *attr.Thing otherwise the
server will crash with a fatal error. Dumping a random memory address is
very ill advised. However when debugging it can be useful to inspect a
Thing even when you cannot reference it by an ALIAS any more - perhaps
it is no longer accessible from where the player is.

Due to its ability to crash the server dumping via a memory address is
only available if the server is running with Server.Debug = true in the
configuration file.

Use with care and use with caution!

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>