attr: Inventory.Move check if a Thing respawnable when removed
When a Thing is removed from an Inventory through either the Move or
Remove methods a check is now made to see if the Thing should be
scheduled to respawn. If it should respawn a non-respawnable copy of the
Thing is returned and the original is scheduled. In this way respawning
items are handled automatically and various commands such as GET and
TAKE do not have to implement their own functionality.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>