attr: Return nothing if Parent of Inventory has a Narrative
At the moment we return "It is empty." for all Inventory attributes that
are empty. That is they have no non-narrative content. This can cause
text to look awkward if the Thing is a narrative (the Thing has a
Narrative attribute). An example is the fish pond. The initial
description is:

  "You examine a pond. This is a small fish pond. There are a few
  aquatic weeds and the water seems quite clear. As you peer in you see
  a few fish quickly dart for the shady cover of the weeds. It contains
  a fish of gold. "

If the "fish of gold" is taken the description becomes:

  "You examine a pond. This is a small fish pond. There are a few
  aquatic weeds and the water seems quite clear. As you peer in you see
  a few fish quickly dart for the shady cover of the weeds. It is
  empty."

Having the "It is empty." on the end looks wrong. For narratives we are
better off not declaring that "It is empty." even when it is.

Another example is the crack in the rock near the shed. Initially the
description is:

  "You examine a small crack in the rock. This is a small crack near the
  base of the rock. It contains a key."

With the key taken we get:

  "You examine a small crack in the rock. This is a small crack near the
  base of the rock. It is empty."

Dropping the "It is empty." we lose nothing - except maybe a hint that
there could be something there?

Last example is the undergrowth in the forest:

  "You examine some undergrowth. This undergrowth contains so many weeds
  and pests it's enough to make a grown gardener cry. It contains a
  small mushroom."

If we take the mushroom the description is again awkward at best:

  "You examine some undergrowth. This undergrowth contains so many weeds
  and pests it's enough to make a grown gardener cry. It is empty."

This change drops the "It is empty." If the parent Thing of the
Inventory has a Narrative attribute.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>