attr: Automatically maintain an alias matching Thing.UID()
Enable a Thing to be uniquely identified via a unique alias that matches
its Thing.UID(). This is so that, as an example, scripted commands such
as $reset and $spawn can operate on a specific thing. Otherwise issuing
a "$RESET RUNE" could pick any thing with a RUNE alias that is available
which could change between the event being queued and the event firing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>