attr: Add automatic Thing cleanup to Inventory.Move
When a Thing is automatically added or moved to an Inventory the Thing
automatically schedules a clean up event if appropriate. It is
appropriate when the Thing does not end up being carried by a player and
is not added or moved to an Inventory that already has a clean up
scheduled.

When a Thing is removed from an Inventory any scheduled cleanups will
automatically be cancelled.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>