cmd: Make output of READ command more generic.
At the moment when you use the READ command it says:

  "You read the writing on <name>. It says: <writing>"

Which is okay for simple items like notes:

  "You read the writing on a note. It says: this is a note."

However when the item is the writing itself the message makes less
sense. For example, assume there are some runes carved into a rock. When
you read the runes the message becomes:

  "You read the writing on some runes. It says: <writing>"

If you are not literally reading the writing it makes even less sense:

  "You read the writing on some runes. It says: The runes depict..."

This change updates the message to a more generic:

  "You read <item>. <writing>"

If zone authors specifically want the "It says:" at the beginning of the
writing it can easily be added per item.

It should also be noted that there are two ways of adding writing to an
item. It can be added directly to the item via a WRITING attribute. The
writing can also be created by adding a separate narrative with a
WRITING attribute that is added to the item's inventory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>