has,attr,cmd,docs: New OnCleanup attribute
OnClean is a simple, single string attribute that can be used to specify
a message to be displayed when an item is cleaned up.

The OnCleanup attribute and interface just provide the basic message.

The $cleanup command has been updated to control when the message is
displayed. If the clean up occurs at a location where it will be seen
then the message will be displayed. If no message has been provided a
default one will be used. If there should be no notification of the
clean up then an OnCleanup attribute with no message will suppress the
default message.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>