has,attr,cmd,docs: New OnReset attribute
OnReset is a simple, single string attribute that can be used to specify
a message to be displayed when an item is reset or respawned.

The OnReset attribute and interface just provide the basic message.

The $reset and $spawn command has been updated to control when the
message is displayed. If the reset or respawn occurs at a location where
it will be seen then the message will be displayed. If no message has
been provided a default one will be used. If there should be no
notification of the reset or respawn then an OnReset attribute with no
message will suppress the default message.

In addition if the reset or respawn occurs within a container and a
message has been provided via an OnReset attribute the message will
propagate to the location. If there is no OnReset the default message
will not be used when the reset or respawn is within a container - i.e.
the reset or respawn will be silent.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>