cmd,attr: Make sure player is never junked + drop exits check
Under no circumstances do we ever want to allow a player to be junked.
At the moment this can happen if a player is put inside a container that
is then junked. The reason being the FindPlayer check is only done for
the specific object being junked and not for contents.

This change adds a JUNK veto to Player. This will trigger automatically
when checking either the player directly or within containers.

In the same check we also test for FindExits. This is not needed as the
JUNK command only looks for aliases in the players Inventory and the
Location inventory - not the Location itself. Thus a location can not be
found to be junked.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>