cmd: Make $ACTION command act on passed alias
Currently $ACTION only acts on the actor that issues the $ACTION
command, that is an item has to issue the command itself. This change
makes the $ACTION command act on the passed alias - which should be a
unique ID alias. This will allow other Thing to issue $ACTION commands
rather than the Thing itself.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>