cmd: Make $CLEANUP command act on passed alias
Currently $CLEANUP only acts on the actor that issues the $CLEANUP
command, that is an item has to issue the command itself. This change
makes the $CLEANUP command act on the passed alias - which should be a
unique ID alias. This will allow other Thing to issue $CLEANUP commands
rather than the Thing itself.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>