cmd: Clean up and improve $RESET & $SPAWN commands
This change documents that we cannot pass an alias to either $RESET or
$SPAWN commands. This is because the actor is the only reference to the
Thing being reset or respawned as it is 'out of play' and not found by
searching Inventories by alias.

It is possible (though currently unlikely) for a player to be inside a
container when an item resets or respawns inside the container with the
player. Currently there is no notification to the player that a reset or
respawn has occurred - messages are propagated up to the location the
inventory is in. This change displays the default - not OnReset -
message to players in the container.

This change also fixes a number of inconsistent variable names.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>