config: Improve debugging configuration
At the moment there is a single Server.Debug flag that has multiple
meanings/uses depending on where it is used. The current uses are:

  - Checking if the #DUMP command can use addresses
  - Checking if a panicking client goroutine should terminate the server

It is desirable to have separate flags that allow specific debugging
features to be enabled or disabled. This change removes the Server.Debug
configuration value and replaces it with a new Debug.<feature> section.
Current debugging features implemented by this change are:

  Debug.Panic - allow client goroutines to panic the server
  Debug.AllowDump - is the #DUMP command allowed

The client handling code, #DUMP command, configuration-file.txt
documentation and data/config.wrj sample configuration have been
updated.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>