cmd: Added #DEBUG command behind Debug.AllowDebug config
This commit adds the #DEBUG command that allows one to:

  - start/stop a memory profile
  - start/stop a CPU profile
  - start/stop a block profile
  - make a heap dump
  - cause a client thread to panic

The memory profile, CPU profile, block profile and heap dumps are
written to the current directory as memprof, cpuprof, blockprof and
heapdump. The *prof files can be inspected with 'go tool pprof'.

The #DEBUG command is only available if Debug.AllowDebug configuration
setting is set to true. This should only be done for development and
debugging.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>