event: Allow panicking event goroutines to be recovered
In the same way that we can recover client goroutines if Debug.Panic is
set to false this commit recovers event goroutines.

If Debug.Panic is set to true the event goroutine will be allowed to
panic which will terminate the server and should only be used for
development and debugging.

NOTE: This feature is already documented in docs/configuration-file.txt.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>