event,config: Add event logging behind Debug.Events config
This commit adds a new configuration option Debug.Events. If set to true
messages for event queuing, delivery and cancellations will be written
to the log. This can cause a lot of messages to be written to the log
which is why the default value for Debug.Events is false in which case
messages for events will not be logged.

The sample server configuration file data/config.wrj and the
documentation in docs/configuration-file.txt have been updated with the
new option.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>