attr,config: Add Thing logging behind Debug.Things config
This commit adds a new configuration option Debug.Things. If set to true
messages for when a Thing is created, discarded or finalized will be
written to the log. This can cause a lot of messages to be written to
the log which is why the default value for Debug.Things is false in
which case messages for Things will not be logged.

The sample server configuration file data/config.wrj and the
documentation in docs/configuration-file.txt have been updated with the
new option.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>