config: Add Debug.LongLog config option + update log format
This commit changes the default log format to be shorter and omits the
file name and line number of where a message is written from:

  2017/04/20 20:04:22 Allocated pool for 624 buffers

It updates the long log format to include times with milliseconds:

  2017/04/20 20:04:22.657371 buffer.go:53: Allocated pool for 624 buffers

If the configuration option Debug.LongLog is set to false (the default)
the shorter format will be used, otherwise the long format will be used.

It should be noted that the server always starts with logging set to the
long format until the configuration has been read and processed, it may
then switch to the shorter format. This is to provide additional
information should there be problems when reading and processing the
configuration file.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>