comms,frontend: Reduce time plain text passwords are vulnerable
Introduction of the #DEBUG command and specifically '#DEBUG heapdump'
has shown that plain text passwords used during login or account
creation can linger in memory for a period of time. It is then possible
to retrieve the passwords from either the heapdumps, core dumps or
memory analysis.

Mitigating this issue has proven to be problematic at best. Memory can
be copied to swap files, held in network buffers, copied during suspend
to disk or copied if running as a virtual machine that is migrated. In
addition there are probably other ways the memory can be copied and
inspected as well.

This change reduces the amount of time that plain text passwords are
held in memory and vulnerable. It is not a 100% foolproof fix for the
security issue. Additional steps would need to be taken outside of
WolfMUD to protect the server process.

One consequence of this change is that the input buffers are of a fixed
size, currently 512bytes. If a user types more than this the input will
be ignored and the message "You type too much." will be displayed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>