recordjar: Fix broken pairs in PairList/KeyedString/KeyedStringList
Pairs of data are expected to an alphanumeric keyword followed by a
non-alphanumeric separator followed by data. For example:

  KEY→data

However the pairs may be malformed with parts missing. For example:

  KEY→
  KEY
  →

In these instances we should attempt to provide a valid pair by
supplying an empty data entry instead of ignoring the pair completely.

Comments updated for PairList/KeyedString/KeyedStringList as required.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>