attr: Protect Locate where & origin fields with RWMutex
In the cmd.state.newState method we need to find the actor's location
and store it in the state struct. However this happens before locking is
established via the state.sync method. This causes a data race. This
change protects the Locate.where and Locate.origin fields with a
RWMutex.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>