attr: Protect Thing attrs field with RWMutex
In the cmd.state.newState method we need to find the actor's location
and store it in the state struct. However this happens before locking is
established via the state.sync method. This causes a data race when
calling Thing.Attrs from within FindLocate. This change protects the
Thing.attrs field with a RWMutex.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>