cmd: Fix default cleanup message not being displayed
When testing if the default clean up message has been supressed we need
to check if an OnCleanup attribute was found with an empty message. It
is not enough to just check if the message is empty as it could also be
empty if no OnCleanup attribute has been specified at all.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>