cmd: Fix scripting of JUNK in cleanup
If clean up will not be seen we should return the s.ok value from the
scripting and not set s.ok to true.

When the clean up will be seen we need to check if JUNK needs to relock.
We also need to script JUNK before send messages to avoid duplicates in
the case of failure or relocking.

NOTE: This fix would be better if instead of having status.ok which is a
boolean we could have a status.result with values for Ok, Fail, Retry
which is expandable in the future. There would then be no need to check
the length of the status.locks slice.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>