cmd: Lock locations as radius expands for SNEEZE command
Currently the SNEEZE command calls Location.Within once and then calls
FindExits on all of the Location found which are not locked.

To make this concurrent safe the Radius passed to Location.Within should
be gradually expanded, re-locking the Location at each increment so that
FindExits is only called on Location that are locked.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>