config: Check file system permissions and set Server.SetPermissions
This commit adds a function, filesystemCheck, to determine whether
filesystem being used supports the setting of permissions. This function
is used to set a new Server configuration variable SetPermissions which
can be used to determine whether we should try setting file permissions
when creating new files such as player account files.

This patch checks the filesystem rather than the OS as it it possible to
have, for example, Linux using a DOS filesystem.

Commit based on bug report from Mikolaj Holysz: bug preventing account
creation on windows due to chmod not being supported.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>