comms: Fix Telnet DEL causing array index out of bounds
On Telnet clients using character at a time processing it is possible to
generate an array index out of bounds error when using the delete key in
some circumstances. This patch reworks the fixDEL function to remove a
lot of corner cases in the processing. It also makes sure that garbage
is not left between the slice length and it's capacity when the slice
shrinks.

Bug reported by Mikolaj Holysz.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>