attr,has: Add disabling/enabling of Thing in Inventory
When a Thing is waiting to reset or respawn it is currently taken out of
play by removing it from all Inventory and hence the world. However this
causes an issue when the reset or respawn occurs and the Thing has to be
put back in to play and hence the world. As the Thing is no longer in an
Inventory it cannot be covered by a lock (BRL - Big Room Lock) which
leads to data races.

This commit adds a way to disable Thing to take them out of play and a
way to enable them again to put them back in play - all while under a
lock (BRL). As a Thing is still associated with an Inventory when
disabled attributes such as Locate still work as expected. However
disabled Thing are separate from an Inventory content and so do not
appear to players as normal items or narratives.

A per Inventory disabled slice was chosen over a single slice so that
normal BRL locking works without any changes. It also avoids contention
with locking a single disabled slice and requires not additional locks
as the Inventory in question would have been locked anyway when adding
the Thing back into the Inventory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>