attr,cmd: Rework resetting/respawning to use Inventory disabling
Currently when a respawn of a Thing is registered a copy is made and put
into play substituting the original Thing. The original Thing is then
removed from the world until the reset triggers. However as the original
Thing is removed from the world there is no way of gaining a BRL (big
room lock) on it when adding it back into the world which leads to data
races.

This commit updates Reset.Spawn so that the original Thing is no longer
removed from the world but disabled and taken out of play instead. That
way the original Thing is always in an Inventory and can be locked under
a BRL. This commit also updates the RESET command to enabled the
original Thing to put it back in play when the reset is triggered.

Respawn has also been updated to call Reset to register the reset event
instead of using it's own respawn event. This means that a respawn and a
reset work the same internally. As a result the RESET command has had
references to spawn dropped.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>