cmd: Rework JUNK command (data races)
Currently, for a Thing that needs to be reset, the JUNK command removes
a Thing from the world so that it is out of play by removing it from all
Inventory. However this means that the Thing cannot be covered by a BRL
(big room lock) when adding it back into the world - putting it back in
to play. This causes data races.

For a Thing that does not to be reset it is removed from all Inventory
in the world and then freed for garbage collection.

This commit uses the new Inventory disabling to take a Thing out of play
that needs to be reset and put back into play.

When junking a Thing any content in Inventory is recursively junked as
well. Previously only the origin of the Thing being junked was locked.
This commit correctly locks the origins of all of the content that is
being junked and reset.

The JUNK command should be the only mechanism for disposing of items in
the world. To work towards this aim JUNK implements it's own dispose
function instead of using Thing.Dispose - which will be removed. This
helps to centralise the mechanism into cmd/junk.go only.

At the moment it is not possible to associate methods with commands as
commands are not types. It has therefore been necessary to add three
temporary helper functions to junk: junkLockAll, junkDispose and
junkCheckVetoes which will be made methods in the future.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>