comms: Rewrite fixDEL to improve performance, handle Unicode
The current version of fixDEL does not handle Unicode then deleting
characters. It's performance is also suboptimal.

A patch was sent in by Mikołaj Hołysz which significantly improved the
performance of fixDEL. However while testing this patch it was
discovered that the old version of fixDEL and the new version could not
handle deleting multi-byte Unicode.

This patch uses a modified version of the algorithm proposed by Mikołaj
and adds Unicode support.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>