cmd: Make command handlers types+methods instead of functions
Currently command handlers for commands such as 'GET' and 'DROP' are
implemented as functions in the package cmd. For example:

  func Get(s *state) { ... }
  func Drop(s *state) { ... }

This makes it very difficult and ugly to implement helper functions or
to split code into multiple functions as there is no 'namespace' to
attach additional functions to. A specific case is the 'JUNK' command
which has implemented the following helpers using the 'junk' prefix as a
pseudo namespace:

  func junkLockAll(s *state, t has.Thing) { ... }
  func junkCheckVetoes(t has.Thing) bool { ... }
  func junkDispose(t has.Thing) { ... }

This change turns the command handlers into types instead of functions
so that methods can be added to the types as appropriate. For example
the handler for 'GET' which is:

  func Get(s *state) { ... }

Becomes:

  type get cmd

  func (get) process(s *state) { ... }

To implement handlers as types instead of functions handler.go has been
updated to register types implementing the handler interface. The method
state.handleCommand has been updated to call the process method defined
by the handler interface. All of the current command functions have been
updated in a similar way to the 'GET' command above.

There is a default command type cmd which is a struct{}, this make the
code more readable:

  type get cmd

Instead of:

  type get struct{}

Although this may change if common helper methods are added to the cmd
type. It should also be noted that the functions, for example Get, were
exported whereas the type get and method get are not. Commands should
not be usable outside of the cmd package however this also prevents the
commands from being seen in godoc - should they be exported just for
documentation purposes?

The version command also renamed the version package level variable to
commit to resolve a duplicate naming conflict.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>