attr,has,cmd: Make exits.Within work as expected
Currently exits.Within does not set the first Inventory in the First
returned slice to be the current location. For example in a container
with no exits the function simply returns. However the current code
cannot determine the current parent and hence the current Inventory if
there are no exits.

This change modifies exits.Within to take an additional parameter which
should be the current location. This allows us to always populate the
current location in the returned slices.

The commands SAY and SNEEZE which use exits.Within have been updated to
pass this new parameter.

Documentation has also been improved.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>