attr,cmd: Move Inventory event removal logic to GET/TAKE commands
Currently the Inventory attribute has too much 'magic' and knows too
much about items being moved from one Inventory to another. An example
is handling events for Thing that are being removed from an Inventory.
As much as possible Attributes should be self contained and it is the
commands that act as the 'glue' to provide functionality. To this end
the logic for when a Thing is removed form a Inventory has been moved to
the GET and TAKE commands. As the logic no longer sits after the Thing
is removed from one Inventory but before being added to the new
Inventory the resetSpawn method has been modified to disable the
original Thing and add+enable the copy produced.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>