attr,has,cmd: Inventory Move/Add/Remove no longer return a Thing
When moving a Thing from one Inventory to another it is no longer
possible that a copy of the Thing will be spawned. Hence we no longer
need to return the Thing that was actually moved. This means we can drop
the returned Thing from the Inventory Move, Add and Remove methods.

Updated has.Inventory interface.

Updated commands that expect Move, Add or Remove to return a Thing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>