attr: Add/Update Locate attribute in AddDisabled
At the moment the Inventory.Move method always checks to see if a Thing
has a Locate attribute. If a Locate attribute is found it is updated,
otherwise a new Locate attribute is added.

Due to new functionality a Thing will only be added to an Inventory via
the Inventory.AddDisabled method. By adding the check and
addition/update of the Locate attribute to AddDisabled it does not have
to be done everytime something moves - only when initially added.

As the Inventory.Move method still performs the checks there is a little
additional overhead until those checks are dropped.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>