attr: Rewrite Inventory.Move slow path without Inventory.Add
Currently the Inventory.Move method uses a call to Inventory.Add for the
slow path when the Inventory to move a Thing to is not a *Inventory but
still implements the has.Inventory interface. This is because the
content is not accessible directly through the interface.

The new slow path now uses a cumbersome disable/remove/add/enable dance
using only methods available in the Inventory interface. Maybe at a
later point this can be improved upon.

Note: The slow path has not been tested as we currently only have the
single interface implementation.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>