attr: Move clean up logic from Inventory.Move to DROP/PUT commands
Currently the Inventory attribute has too much 'magic' and knows too
much about items being moved from one Inventory to another. An example
is handling events for Thing that is being added to an Inventory.
As much as possible Attributes should be self contained and it is the
commands that act as the 'glue' to provide functionality. To this end
the logic for when to register a Thing for clean up has been moved to
the DROP and PUT commands.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>