attr: Move action logic from Inventory.Move to DROP command
Currently the Inventory attribute has too much 'magic' and knows too
much about items being moved from one Inventory to another. An example
is handling events for Thing that is being added to an Inventory. As
much as possible Attributes should be self contained and it is the
commands that act as the 'glue' to provide functionality. To this end
the logic for when to register a Thing for actions has been moved to the
DROP command. The zone loader has also been updated In order to have
actions initially active by default.

At this point in time actions are not re-enabled for a Thing PUT into a
container. The only reason to prevent this at the moment is the fact that
action messages may 'look odd' expecting a Thing to be 'in the open'. It
may be necessary to have two sets of actions 'open' and 'confined' for
when a Thing is either in the open and free or confined in a container.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>