attr: Reclaim empty slice storage in Inventory.RemoveDisabled
If the disabled list for an Inventory is empty and it has a capacity
create a new empty disabled list with zero capacity. This can save on
memory usage at the expense of an additional slice to garbage collect.
We check the capacity to avoid recreating zero lenth slices that have
not been used.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>