recordjar: Rework encoder/decoder into own sub packages
At the moment using the recordjar encoder and decoder is quite verbose:

  recordjar.Decode.Duration(data)
  recordjar.Encode.Duration(data)

There is also the issue of only exporting functionality needed by other
packages. However this means that the useful functions are not visible
by default in the go doc.

This change moves the encoder and decoder into their own packages. It
drops the exported Decode and Encode package level variables. The
unexported decode and encode types are dropped and their methods turned
into functions. This allows for less verbose usage of the encoder and
decoder:

  decode.Duration(data)
  encode.Duration(data)

The functions for both the decoder and encoder are exported and visible
in the go doc.

Usage of the decoder and encoder in the attr, frontend and zones
packages have been updated to use the new packages and functions.

Suggested-by: Mikolaj Holysz <miki123211@gmail.com>
Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>