attr,cmd: Fix spawn check in Reset and usage in GET/TAKE
There is currently an issue where an Thing without a Reset attribute
seems to be taken, but it is not actually put into the players
inventory.

If a Thing does not have a Reset attribute then the original Thing can
not be returned properly by the Spawn method as we cannot get the Reset
attribute's parent Thing. This results in a nil pointer which the Move
method on an Inventory will silently ignore.

This change modifies Spawn to return nil, instead of the original Thing,
when a Thing does not respawn for any reason. The commands GET and TAKE
have been modified to only assign the result of Spawn if it is not nil.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>