attr,has: Add string tag return parameter to Marshal interface
When marshaling attributes the tag used for the attribute is currently
unknown. For example "alias" or "aliases" for the type Alias attribute.
As in this example the tag may depend on the data returned by Marshal.

This change adds a string tag return parameter to the Marshaler
interface definition of Marshal. This allows Marshal to return an
appropriate tag for the data marshaled. The tag is then used in a
recordjar record as the field name.

This change also updates the temporary implementation of the Attribute
Marshal method.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>