recordjar: decode.PairList now returns map[string]string
The decode.PairList function currently returns a slice of string arrays
each two elements in length. This leads to the elements being accessed
using 'magic' numbers such as pair[0] and pair[1] when working with the
returned values. Accessing the pairs also requires looping over the
slice to retrieve the arrays, and then accessing the array elements.

This change replaces [][2]string with a map[string]string which is
simpler to range over and requires no 'magic' numbers in the code.

The only change in functionality is that the map cannot have duplicate
keys. This means that a duplicate name on the left of a pair will
overwrite a previous value:

    Exits: E→L1 E→L2

In this example "E" would have the value "L2". As duplicate names are
not currently handled or expected this is not an issue.

This change also updates attributes and the zone loader to use the new
returned type.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>